+++ lightly_studio_view/src/lib/components/TagCreateDialog/TagCreateDialog.svelte
@@ -267,6 +267,7 @@
         hasFetched = false;
         tagsEnlistedToCreate = [];
         tagsToAddItemsTo.clear();
+        itemsSelectedByFilter = null;
     };
 
     const changesToCommit = $derived(tagsToAddItemsTo.size > 0 || tagsEnlistedToCreate.length > 0);