+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added VideoSampleField for querying video datasets. VideoDataset.query() now works.
 - Added helper functions and a tutorial on running Python and the GUI in parallel.
 - Added image sample loader from Lightly prediction format.
+- Added image classification editing: users can now add/remove/change image classification.
 
 ### Changed